remove required math classes page
Submitted an issue about this topic a couple months back -- I'll summarize here:

This page is not up-to-date. Some of the information on it is factually incorrect (e.g. A&S students do not need 1920 to affiliate with the CS major), and some sections are just empty. Rather than trying to keep this page up-to-date, I think we would be better off linking students to the official CS major page. This way we wouldn't be violating the DRY principle, and we wouldn't have to worry about updating incorrect information.