Remove dead links to CS courses without pages
Rather than have some classes that are listed
but don't have a page and others that aren't
listed at all, all CS courses without a page
have been removed. The course roster has a
list of courses, so we do not need to fill
that role. Having dead links also discourages
users and gives the appearance of an
abandoned site.

Note: CS 2802 was kept (the page will be added
shortly).

A small formatting error with CS 4160 was
fixed as well.